# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: Portfolio Analysis API
  description: |
    Portfolio analysis backend. Parses InteractiveBrokers FlexQuery XML reports,
    fetches historical market data from Yahoo Finance, and provides portfolio
    valuation, statistics (MWR/TWR), and benchmark comparisons.

    ## Authentication
    Every request must include an `X-Auth-Token` header containing the user's
    pre-shared secret token.  The server hashes the token (SHA-256) to derive a
    stable user identifier used for data isolation on disk.

    ## Accounting Models
    Several endpoints accept an `accounting_model` parameter that controls how
    multi-currency positions are valued:

    | Model | Description |
    |-------|-------------|
    | `historical` | Each position is converted to the target currency using the FX rate on the date the value is computed (i.e. each day uses that day's rate). **This is the default.** |
    | `spot` | All positions are converted using today's (latest available) FX rate, regardless of the valuation date. Useful for quick snapshots but distorts historical series. |
    | `original` | No currency conversion is performed; values are returned in each security's native currency. The response may contain mixed currencies. |
  version: "1.0.0"

servers:
  - url: http://localhost:8080

security:
  - tokenAuth: []

components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-Auth-Token

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    PortfolioValueResponse:
      type: object
      properties:
        values:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Map of currency code → portfolio value, e.g. {\"USD\": 12345.67, \"EUR\": 11000.00, \"CZK\": 280000.00}"
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionValue"

    PositionValue:
      type: object
      properties:
        symbol:
          type: string
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
          description: Latest price in the security's native currency
        native_currency:
          type: string
        cost_basis:
          type: number
          format: double
          description: Average cost basis per share in the security's native currency
        values:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Position value in each requested currency"

    TradeEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        side:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
          description: Trade price in the security's native currency
        native_currency:
          type: string
        converted_price:
          type: number
          format: double
          description: Trade price converted to the requested display currency
        commission:
          type: number
          format: double
        proceeds:
          type: number
          format: double

    TradesResponse:
      type: object
      properties:
        symbol:
          type: string
        currency:
          type: string
          description: Security's native currency
        display_currency:
          type: string
          description: Currency used for converted_price
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeEntry"

    PortfolioHistoryResponse:
      type: object
      properties:
        currency:
          type: string
        accounting_model:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/DailyValue"

    DailyValue:
      type: object
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          format: double

    MarketHistoryResponse:
      type: object
      properties:
        symbol:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"

    PricePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        open:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        close:
          type: number
          format: double
        adj_close:
          type: number
          format: double
        volume:
          type: integer
          format: int64

    StatsResponse:
      type: object
      properties:
        currency:
          type: string
        accounting_model:
          type: string
        statistics:
          type: object
          additionalProperties: true
          description: |
            Map of statistic name → computed value.  Always includes:
            - `mwr` (Money-Weighted Return / IRR)
            - `twr` (Time-Weighted Return)
            Additional registered calculators appear here automatically.

    CompareResponse:
      type: object
      properties:
        currency:
          type: string
        accounting_model:
          type: string
        benchmarks:
          type: array
          items:
            $ref: "#/components/schemas/BenchmarkResult"

    BenchmarkResult:
      type: object
      properties:
        symbol:
          type: string
        alpha:
          type: number
          format: double
        beta:
          type: number
          format: double
        sharpe_ratio:
          type: number
          format: double
        treynor_ratio:
          type: number
          format: double
        tracking_error:
          type: number
          format: double
        information_ratio:
          type: number
          format: double
        correlation:
          type: number
          format: double

    UploadResponse:
      type: object
      properties:
        message:
          type: string
        positions_count:
          type: integer
        trades_count:
          type: integer
        cash_transactions_count:
          type: integer

paths:
  /api/v1/portfolio/upload:
    post:
      operationId: uploadFlexQuery
      summary: Upload InteractiveBrokers FlexQuery XML
      description: |
        Upload an IB FlexQuery Statement XML file. The report is parsed and
        persisted to disk under the user's hashed identifier. Subsequent
        portfolio queries operate on this data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: FlexQuery XML file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Bad request (missing file, parse error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/portfolio/value:
    get:
      operationId: getPortfolioValue
      summary: Get current portfolio value
      description: |
        Returns the current total portfolio value converted to the requested
        currencies (default: USD, EUR, CZK).  Also returns a per-position
        breakdown.
      parameters:
        - name: currencies
          in: query
          description: Comma-separated currency codes (default "USD,EUR,CZK")
          schema:
            type: string
            default: "USD,EUR,CZK"
      responses:
        "200":
          description: Current portfolio value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioValueResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No portfolio uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/portfolio/trades:
    get:
      operationId: getPortfolioTrades
      summary: Get trade history for a symbol
      description: |
        Returns all trades for a specific symbol from the uploaded FlexQuery
        report. Prices are shown in both the security's native currency and
        a display currency of the user's choice.
      parameters:
        - name: symbol
          in: query
          required: true
          description: Ticker symbol to retrieve trades for
          schema:
            type: string
        - name: currency
          in: query
          description: Display currency for converted prices (default CZK)
          schema:
            type: string
            default: CZK
      responses:
        "200":
          description: Trade history for the symbol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradesResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No portfolio uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/portfolio/history:
    get:
      operationId: getPortfolioHistory
      summary: Get daily portfolio value over time
      description: |
        Returns the portfolio value for each day in the given range, converted
        to the specified currency using the chosen accounting model.
      parameters:
        - name: from
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: currency
          in: query
          required: true
          description: Target currency code (e.g. USD, EUR, CZK)
          schema:
            type: string
        - name: accounting_model
          in: query
          description: |
            Currency conversion model: `historical` (default), `spot`, or `original`.
            See the top-level description for details.
          schema:
            type: string
            default: historical
            enum: [historical, spot, original]
      responses:
        "200":
          description: Daily portfolio value series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHistoryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No portfolio uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/market/history:
    get:
      operationId: getMarketHistory
      summary: Get historical price data for a symbol
      description: |
        Returns OHLCV data for the given stock/ETF/index symbol over the
        specified date range.  Data is fetched from Yahoo Finance and cached
        on disk.
      parameters:
        - name: symbol
          in: query
          required: true
          description: Ticker symbol (e.g. AAPL, VWCE.DE, ^GSPC)
          schema:
            type: string
        - name: from
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Historical price data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketHistoryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/portfolio/stats:
    get:
      operationId: getPortfolioStats
      summary: Get portfolio statistics
      description: |
        Returns performance statistics for the current portfolio.  All
        calculations are performed in the specified currency using the chosen
        accounting model.

        **Included statistics:**
        - `mwr` — Money-Weighted Return (IRR), solved via Newton's method.
        - `twr` — Time-Weighted Return, chaining sub-period returns between
          external cash-flow events.

        The system is extensible: additional stat calculators can be registered
        and will appear here automatically.
      parameters:
        - name: currency
          in: query
          required: true
          description: Currency in which to compute statistics
          schema:
            type: string
        - name: accounting_model
          in: query
          description: |
            Currency conversion model: `historical` (default), `spot`, or `original`.
          schema:
            type: string
            default: historical
            enum: [historical, spot, original]
      responses:
        "200":
          description: Portfolio statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No portfolio uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/portfolio/compare:
    get:
      operationId: comparePortfolio
      summary: Compare portfolio against benchmarks
      description: |
        Compares the current portfolio's performance against one or more
        benchmark symbols (stocks, ETFs, indices).  Returns risk-adjusted
        metrics computed in the given currency.

        All return series are first converted to the target currency
        using the selected accounting model before comparison.
      parameters:
        - name: symbols
          in: query
          required: true
          description: Comma-separated benchmark symbols (e.g. "SPY,VWCE.DE")
          schema:
            type: string
        - name: currency
          in: query
          required: true
          description: Currency in which to compute comparison
          schema:
            type: string
        - name: from
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: accounting_model
          in: query
          description: |
            Currency conversion model: `historical` (default), `spot`, or `original`.
          schema:
            type: string
            default: historical
            enum: [historical, spot, original]
        - name: risk_free_rate
          in: query
          description: Annual risk-free rate for Sharpe/Treynor (default 0.05 = 5%)
          schema:
            type: number
            format: double
            default: 0.05
      responses:
        "200":
          description: Benchmark comparison results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No portfolio uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"